Simplify the code
