+++ config/src/open/mod.rs
@@ -1,5 +1,7 @@
 mod open;
 mod opener;
+mod rule;
 
 pub use open::*;
 pub use opener::*;
+use rule::*;
