+++ config/src/open/open.rs
@@ -1,25 +1,16 @@
-use std::{collections::BTreeMap, fmt, path::Path};
+use std::{collections::BTreeMap, path::Path};
 
 use indexmap::IndexSet;
-use serde::{de::{self, Visitor}, Deserialize, Deserializer};
+use serde::{Deserialize, Deserializer};
 use shared::MIME_DIR;
 
 use super::Opener;
-use crate::{Pattern, MERGED_YAZI};
+use crate::{open::OpenRule, MERGED_YAZI};
 
 #[derive(Debug)]
 pub struct Open {
 	openers: BTreeMap<String, IndexSet<Opener>>,
 	rules:   Vec<OpenRule>,
-}
-
-#[derive(Debug, Deserialize)]
-struct OpenRule {
-	name: Option<Pattern>,
-	mime: Option<Pattern>,
-	#[serde(rename = "use")]
-	#[serde(deserialize_with = "deserialize_from_str_or_vec")]
-	use_: Vec<String>,
 }
 
 impl Default for Open {
@@ -40,15 +31,11 @@ impl Open {
 				let openers = rule
 					.use_
 					.iter()
-					.filter_map(|use_name| self.openers.get(use_name))
+					.filter_map(|use_| self.openers.get(use_))
 					.flatten()
 					.collect::<IndexSet<_>>();
 
-				if openers.is_empty() {
-					return None;
-				}
-
-				Some(openers)
+				if openers.is_empty() { None } else { Some(openers) }
 			} else {
 				None
 			}
@@ -61,13 +48,13 @@ impl Open {
 		P: AsRef<Path>,
 		M: AsRef<str>,
 	{
-		self.openers(path, mime).and_then(|o| o.iter().find(|o| o.block).copied())
+		self.openers(path, mime).and_then(|o| o.into_iter().find(|o| o.block))
 	}
 
 	pub fn common_openers(&self, targets: &[(impl AsRef<Path>, impl AsRef<str>)]) -> Vec<&Opener> {
-		let grouped = targets.iter().filter_map(|(p, m)| self.openers(p, m)).collect::<Vec<_>>();
-		let flat = grouped.iter().flatten().collect::<IndexSet<_>>();
-		flat.into_iter().filter(|&o| grouped.iter().all(|g| g.contains(o))).copied().collect()
+		let grouped: Vec<_> = targets.iter().filter_map(|(p, m)| self.openers(p, m)).collect();
+		let flat: IndexSet<_> = grouped.iter().flatten().copied().collect();
+		flat.into_iter().filter(|&o| grouped.iter().all(|g| g.contains(o))).collect()
 	}
 }
 
@@ -90,46 +77,4 @@ impl<'de> Deserialize<'de> for Open {
 		let openers = outer.opener.into_iter().map(|(k, v)| (k, IndexSet::from_iter(v))).collect();
 		Ok(Self { openers, rules: outer.open.rules })
 	}
-}
-
-fn deserialize_from_str_or_vec<'de, D>(deserializer: D) -> Result<Vec<String>, D::Error>
-where
-	D: Deserializer<'de>,
-{
-	struct StringVisitor;
-
-	impl<'de> Visitor<'de> for StringVisitor {
-		type Value = Vec<String>;
-
-		fn expecting(&self, formatter: &mut fmt::Formatter) -> fmt::Result {
-			formatter.write_str("a string, or array of strings")
-		}
-
-		fn visit_seq<A>(self, mut seq: A) -> Result<Self::Value, A::Error>
-		where
-			A: de::SeqAccess<'de>,
-		{
-			let mut strs = Vec::new();
-			while let Some(value) = seq.next_element::<String>()? {
-				strs.push(value);
-			}
-			Ok(strs)
-		}
-
-		fn visit_str<E>(self, value: &str) -> Result<Self::Value, E>
-		where
-			E: de::Error,
-		{
-			Ok(vec![value.to_owned()])
-		}
-
-		fn visit_string<E>(self, v: String) -> Result<Self::Value, E>
-		where
-			E: de::Error,
-		{
-			Ok(vec![v])
-		}
-	}
-
-	deserializer.deserialize_any(StringVisitor)
 }
